work around a lockdown browser rendering issue
lockdown browser has an issue when the base explorer install is IE9 where the
first page rendered after a popup window is opened does not render correctly.
since it only happens in the first page after the popup is created, we are
working around it for now by loading a dummy page as the first page which
automatically redirects to the actual quiz page.

fixes CNVS-3881
refs CNVS-4853

test plan
- make sure taking quizzes works fine with LDB
- in a system with LDB and a base IE9 installed
- start taking a quiz, it should render correctly
- try with/without oqaat, and other quiz options

Reviewed-on: https://gerrit.instructure.com/19272
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Product-Review: Bryan Madsen <bryan@instructure.com>

+++ app/controllers/quizzes_controller.rb
@@ -155,6 +155,10 @@ class QuizzesController < ApplicationController
         return unless check_lockdown_browser(:medium, named_context_url(@context, 'context_quiz_url', @quiz.to_param, :viewing => "1"))
       end
 
+      if @quiz.require_lockdown_browser? && refresh_ldb = value_to_boolean(params.delete(:refresh_ldb))
+        return render(:action => "refresh_quiz_after_popup")
+      end
+
       @question_count = @quiz.question_count
       if session[:quiz_id] == @quiz.id && !request.xhr?
         session.delete(:quiz_id)
